[refactor] 근속 전망 통계 테이블 insert 문제 수정

+++ momentum-dao-be/src/main/java/com/dao/momentum/organization/employee/command/infrastructure/repository/JpaEmployeeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface JpaEmployeeRepository extends EmployeeRepository, JpaRepository<Employee, Integer> {
+public interface JpaEmployeeRepository extends EmployeeRepository, JpaRepository<Employee, Long> {
     <S extends Employee> S save(Employee employee);
 
     Optional<Employee> findByEmail(String email);
@@ -26,5 +26,6 @@ public interface JpaEmployeeRepository extends EmployeeRepository, JpaRepository
 
     // 근속 전망 확인을 위한 현 재직 중인 사원 목록 조회
     List<Employee> findByStatus(String status); // Spring Data JPA 규칙 기반
+
 
 }
